requirements.txt
requests>=2.31.0
pyyaml>=6.0.1
twilio>=8.0.0
